refactor codes and continue to change directory structures
