+++ README.md
@@ -1,2 +1,12 @@
 ## Description
 for GCP cost monitoring
+
+
+## Setup
+export BILLING_PROJECT_ID="YOUR_PROJECT_ID"
+export BILLING_BQ_TBL="YOUR_BILLING_TABLE"
+export CREDENTIAL_PATH="YOUR_CLIENT_SECRET_PATH"
+export GSS_ID="GSS_ID"
+export TOKEN_PATH="ACCESS_TOKEN_PATH"
+
+need to run before deploy to cloud functions
