+++ cloudfunctions/monthly_billing/.gitignore
@@ -1,0 +1,1 @@
+token.json
